requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
datasets==4.3.0
dill==0.4.0
distro==1.9.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
idna==3.11
jiter==0.11.1
multidict==6.7.0
multiprocess==0.70.16
networkx==3.5
numpy==2.3.4
openai==2.7.0
packaging==25.0
pandas==2.3.3
propcache==0.4.1
pyarrow==22.0.0
pydantic==2.12.3
pydantic_core==2.41.4
pydot==4.0.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tiktoken==0.12.0
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
jinja2>=3.0.0
docker~=7.1.0
mini-swe-agent==1.17.4
